refactor: Restruct Mapper Directory to Seperate domain
